Move fake clients to a different file
